third_party/WebKit/LayoutTests/fast/canvas/canvas-font-min.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<canvas width="100" height="100"></canvas>
<script>
test(function() {
  var ctx = document.querySelector('canvas').getContext('2d');
  // Set same font as will be computed on the canvas element.
  ctx.font = "1px sans-serif";
  assert_equals(ctx.font, "1px sans-serif");
}, "Minimum font size not enforced on Canvas.");
</script>
